requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
configparser==7.1.0
    # via -r requirements.in
google-api-core==2.21.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via -r requirements.in
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via google-api-core
idna==3.10
    # via requests
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pymysql==1.1.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
urllib3==2.2.3
    # via requests
